requirements.txt
gradio>=5.0.0
openai>=1.50.0
pypdf>=5.0.0
python-dotenv>=1.0.0
huggingface-hub<1.0.0
chromadb>=0.5.0
pydantic>=2.0.0